+++ src/main/java/com/knowway/admin/service/AdminService.java
@@ -5,6 +5,6 @@ import com.knowway.admin.dto.AdminRecordResponse;
 import java.util.List;
 
 public interface AdminService {
-    List<AdminRecordResponse> getRecordsByFloorId(Integer floorId);
+    List<AdminRecordResponse> getRecordsByFloorId(Long floorId);
     boolean toggleRecordIsSelected(Long id);
 }
